Wow, found why letters were ending up running into each other. The cache for char widths was an array of uchar, which broke as soon as a char width was >256 (which is almost always the case in layout units!) Using ushort instead. Obviously, makes little sense to backport to the real QRT...
svn path=/trunk/koffice/; revision=111596

+++ lib/kotext/qrichtext_p.h
@@ -1483,7 +1483,8 @@ private:
     uint linkColor : 1;
     int leftBearing, rightBearing;
     VerticalAlignment ha;
-    uchar widths[ 256 ];
+    //uchar widths[ 256 ];
+    ushort widths[ 256 ];
     int hei, asc, dsc;
     QTextFormatCollection *collection;
     int ref;
